Update to use RouteResult(Subject|Observer)Interface from zend-expressive-router
Updates `UrlHelper` to use the `RouteResultObserverInterface` from the
zend-expressive-router package instead of from the zend-expressive package.
Additionally, updates the `UrlHelperFactory` tests to generate a mock of
`RouteResultSubjectInterface` to use for testing the "Application" injection.

The tests still specify `Zend\Expressive\Application` in the import statements
and when specifying names to the container mock instance. This works, even
without the class, due to the fact that PHP resolves the `::class` constant
based on the import statements when not operating on an instance.