Added strategies to the BodyParamsMiddleware
This patch builds on the one in #3, moving the act of parsing a Content-Type as
well as parsing the request body to *strategies*, which implement
`Zend\Expressive\BodyParams\StrategyInterface`. These define two methods:

- `match($contentType)`, which should accept a Content-Type header value, and
  return a boolean value indicating whether or not the strategy matches.
- `parse(ServerRequestInterface $request)`, which should accept the request
  instance, parse its raw body, and return a new request instance with the
  results of parsing injected as the parsed body.

The form and json matching/parsing were rewritten as strategies, and added as
default strategies to the middleware. Custom strategies can be written and then
attached using `addStrategy()`.