Ensure that generate() default arguments match those of __invoke()
Prior to this patch, `generate()` was setting the default argument for
the `$fragment` argument to `''`, while `__invoke()` was setting it to
`null`. The issue is that `''` is considered an invalid fragment, and
`__invoke()` specifically tests for a `null` value before validating a
fragment.

This patch updates the `generate()` signature to match that of
`__invoke()`, solving the issue.