Test to cover zendframework/zend-expressive#371
Creates an integration test to attempt to recreate the problem reported
in zendframework/zend-expressive#371. It sets up the following
conditions:

- A request method that can provide body content
- A request body with JSON content
- A `Content-Type` request header with the value `application/json;charset=utf-8`

It then passes the request to the `BodyParams` middleware, and validates
that the request received by the delegate has the parsed contents.

Unfortunately... it does not validate the issue reported, as the test
passes.