refactor: make TemplateVariableContainer immutable
Per a suggestion from @vaclavvanik, I have refactored the
TemplateVariableContainer to be immutable. `set()` was renamed to
`with()`, and `unset()` to `without()`. Each of these methods, as well
as `merge()`, now return a new instance cloned from the previous, with
the changes requested.

This also meant that the RouteTemplateVariableMiddleware needed updates
in order to reset the `TemplateVariableContainer` request attribute.